moved logic for UpdateUser endpoint to UserEndpointsLogic.cs. Tested add token in js (worked, commented out). Some small syntax fixes in rooms.html.
