+++ frontend/js/scripts.js
@@ -48,8 +48,16 @@ async function fetchUsers(resetPage = false, filters = {}) {
         }
     });
 
+    /* let token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJmNDdhYzEwYi01OGNjLTQzNzItYTU2Ny0wZTAyYjJjM2Q0NzEiLCJuYW1lIjoic2FyYWhAZXhhbXBsZS5jb20iLCJwYXNzd29yZFVwZGF0ZWQiOiJ0cnVlIiwiaHR0cDovL3NjaGVtYXMubWljcm9zb2Z0LmNvbS93cy8yMDA4LzA2L2lkZW50aXR5L2NsYWltcy9yb2xlIjpbIkFkbWluIiwiVXNlciJdLCJleHAiOjE3Mzk1NDgwMTcsImlzcyI6InJvb21TY2hlZHVsZXJfQVBJIiwiYXVkIjoicm9vbVNjaGVkdWxlcl9BUEkifQ.1jVol6_oBRowAUJIFSUqeQfhHitryCEA4hp4XCy-OXY";
+    // Build the fetch options object with headers
+    const fetchOptions = {
+        headers: {
+            'Authorization': `Bearer ${token}`
+        }
+    }; */
+
     try {
-        const response = await fetch(url);
+        const response = await fetch(url/*, fetchOptions */);
         if (response.status === 404) {
             currentPage = Math.max(1, currentPage - 1);
             document.getElementById('nextPageButton').disabled = true;
