+++ frontend/rooms.html
@@ -14,7 +14,7 @@
             <img src="images/getRoom_clean.png" alt="Company Logo" id="logo"/>
         </a>   
     </header>
-    <h1>Rooms Managment</h1>
+    <h1>Rooms Management</h1>
     <div id = "adminPanel" style="display: none;">
         <div class = "nav-links">       
             <button id="manageUsersButton">Manage Users</button>
