+++ src/Core/Extensions/WebAppExtensions.cs
@@ -32,7 +32,7 @@ public static class WebAppExtensions
                 builder =>
                 {
                     builder
-                        // .WithOrigins("http://127.0.0.1:56024", "http://localhost:56024") // if frontend only
+                        // .WithOrigins("http://127.0.0.1:56024", "http://localhost:56024") 
                         .AllowAnyOrigin()
                         .AllowAnyHeader()
                         .AllowAnyMethod();
